refactor: change SyncJ number to int

+++ cmd/helper.go
@@ -50,7 +50,7 @@ func isUserError(err error) bool {
 	return userErrorRegexp.MatchString(err.Error())
 }
 
-func min(args ...uint64) uint64 {
+func min(args ...int) int {
 	m := args[0]
 	for _, arg := range args[1:] {
 		if arg < m {
